requirements.txt
Babel==2.5.1
beautifulsoup4==4.6.3
Django==1.11.1
django-appconf==1.0.2
django-axes==4.4.0
django-ckeditor==5.5.0
django-ipware==2.1.0
django-js-asset==1.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
docxtpl==0.5.5
Jinja2==2.9.6
lxml==4.1.0
MarkupSafe==1.0
Pillow==5.1.0
psycopg2==2.7.4
python-docx==0.8.6
pytz==2017.2
six==1.11.0
